backend/requirements.txt
fastapi[all] 
uvicorn
pydantic
sqlalchemy
psycopg2-binary
watchfiles
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
orjson
httpx
pgvector
# ✅ CrewAI 관련 패키지
crewai
exa_py

# ✅ LangChain 및 벡터 DB 관련
langchain-huggingface
langchain-core
langchain
langchain-community
# ✅ NLP 및 모델 관련
transformers
pandas
#
faiss-cpu
torch
kiwipiepy